+++ lib/src/wire/builders/widget_builders/scaffold_builder.dart
@@ -8,16 +8,27 @@ class ScaffoldBuilder extends WidgetBuilder {
     return material.Scaffold(
       appBar: element.data["appBar"] != null
           ? application
-          .make<PreferredSizeWidgetBuilder>(element.data["appBar"]["type"])
-          .build(Element.fromJson(element.data["appBar"]))
+              .make<PreferredSizeWidgetBuilder>(element.data["appBar"]["type"])
+              .build(Element.fromJson(element.data["appBar"]))
           : null,
       body: application
           .make<WidgetBuilder>(element.data["body"]["type"])
           .build(Element.fromJson(element.data["body"])),
+      floatingActionButtonLocation:
+          element.data["floatingActionButtonLocation"] == null
+              ? material.FloatingActionButtonLocation.centerDocked
+              : FloatingActionButtonLocation.fromJson(
+                      element.data["floatingActionButtonLocation"])
+                  .build(),
+      floatingActionButton: element.data["floatingActionButton"] != null
+          ? application
+              .make<WidgetBuilder>(element.data["floatingActionButton"]["type"])
+              .build(Element.fromJson(element.data["floatingActionButton"]))
+          : null,
       bottomNavigationBar: element.data["bottomNavigationBar"] != null
           ? application
-          .make<WidgetBuilder>(element.data["bottomNavigationBar"]["type"])
-          .build(Element.fromJson(element.data["bottomNavigationBar"]))
+              .make<WidgetBuilder>(element.data["bottomNavigationBar"]["type"])
+              .build(Element.fromJson(element.data["bottomNavigationBar"]))
           : null,
     );
   }